CAUSEWAY-3486: further remove debug code
