+++ persistence/jdo/datanucleus/src/main/java/org/apache/causeway/persistence/jdo/datanucleus/entities/DnStateManagerForHollow.java
@@ -18,6 +18,8 @@
  */
 package org.apache.causeway.persistence.jdo.datanucleus.entities;
 
+import java.util.Optional;
+
 import org.datanucleus.ExecutionContext;
 import org.datanucleus.FetchPlan;
 import org.datanucleus.FetchPlanForClass;
@@ -34,8 +36,10 @@ import org.datanucleus.store.FieldValues;
 import org.datanucleus.store.StoreManager;
 import org.datanucleus.store.fieldmanager.FieldManager;
 import org.datanucleus.transaction.Transaction;
+import org.springframework.lang.Nullable;
 
 import lombok.NonNull;
+import lombok.val;
 
 /**
  * Used to memoize stringified OIDs of entities,
@@ -44,6 +48,21 @@ import lombok.NonNull;
  * Has no purpose otherwise.
  */
 public class DnStateManagerForHollow implements DNStateManager<Persistable> {
+
+    // -- UTILITTY
+
+    public static Optional<String> lookupIdentifierFor(final @Nullable Object pojo) {
+        if(pojo instanceof Persistable) {
+            var sm = ((Persistable) pojo).dnGetStateManager();
+            if(sm instanceof DnStateManagerForHollow) {
+                val oidStringified = ((DnStateManagerForHollow)sm).oidStringified;
+                return Optional.of(oidStringified);
+            }
+        }
+        return Optional.empty();
+    }
+
+    // -- CONSTRUCTION
 
     public final @NonNull String oidStringified;
